--- .eslintrc.yml
@@ -1,7 +1,0 @@
-env:
-  browser: true
-  es2021: true
-extends: eslint:recommended
-parserOptions:
-  ecmaVersion: latest
-rules: {}
