+++ index.html
@@ -50,37 +50,37 @@
                     <img src='./images/Enabled.png' alt='Menu' class='menu' onclick='toogle()' />
                 </nav>
             </div>
-            <div class="cont " id="nav ">
-                <nav class="nav-item1 ">
+            <div class="cont">
+                <nav class="nav-item1">
                     <a href=" ">Portfolio</a>
                 </nav>
-                <nav class="nav-item1 ">
+                <nav class="nav-item1">
                     <a href=" ">About</a>
                 </nav>
-                <nav class="nav-item1 ">
+                <nav class="nav-item1">
                     <a href=" ">Contact</a>
                 </nav>
             </div>
         </header>
-        <section class="content ">
-            <h1 class="box1 ">
+        <section class="content">
+            <h1 class="box1">
                 Hey there. I’m Alex.
                 <span>I’m a software developer</span>
             </h1>
-            <p class="box2 ">
+            <p class="box2">
                 I can help you build a product , feature or website Look through some of my work and experience! If you like what you see and have a project you need coded, don’t hestiate to contact me.
             </p>
-            <ul class="box3 ">
+            <ul class="box3">
                 <li class="social-media ">
                     <a href=" "><img src="images/twitter.png " alt="twitter " /></a>
                 </li>
-                <li class="social-media ">
+                <li class="social-media">
                     <a href=" "><img src="images/Linkedin.png " alt="Linkedln " /></a>
                 </li>
-                <li class="social-media ">
+                <li class="social-media">
                     <a href=" "><img src="images/github.png " alt="m " /></a>
                 </li>
-                <li class="social-media ">
+                <li class="social-media">
                     <a href=" "><img src="images/fingers.png " alt="github " /></a>
                 </li>
                 <li class="social-media ">
@@ -92,13 +92,12 @@
         </button>
         </section>
     </div>
-    </header>
     <section class="work-section ">
         <h2 class="work ">
             My Recent Works
         </h2>
         <p><img src="images/Indicator.png " alt="line " class="line " /></p>
-        <section id ="My-works" class="My-Works">
+        <section id="My-works" class="My-Works">
         </section>
     </section>
     <section class="about-me " id='about'>
@@ -114,7 +113,7 @@
 
         <section class="land ">
             <section class="language third ">
-                <h2 class="titre " id="fonct ">Languages</h2>
+                <h2 class="titre " id="fonct">Languages</h2>
                 <ul class="list ">
                     <li class="lan ">Javascript</li>
                     <li class="lan ">Ruby</li>
@@ -144,54 +143,54 @@
             </section>
         </section>
     </section>
-    <footer class="contact-form " id='contact'>
-        <div class="desktop ">
+    <footer class="contact-form" id='contact'>
+        <div class="desktop">
             <p class="contact-paragraph ">I'm always interested in hearing about new projects, so if you'd like to chat please get in touch.
             </p>
 
-            <form action="https://formspree.io/f/xqknqpzn " method="post ">
-                <ul class="contact ">
-                    <li class="contact-list ">
-                        <label for="name "></label>
-                        <input type="text " id="name " name="user_name " placeholder="Full Name (max 30 characters) " class="contact-input " maxlength="30 " required />
+            <form action="https://formspree.io/f/xqknqpzn" method="post">
+                <ul class="contact">
+                    <li class="contact-list">
+                        <label for="name"></label>
+                        <input type="text" id="name" name="user_name" placeholder="Full Name (max 30 characters)" class="contact-input " maxlength="30" required />
                     </li>
                     <li class="contact-list ">
-                        <label for="mail "></label>
-                        <input type="email " id="mail " name="user_email " placeholder="Email address " class="contact-input " pattern="[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,4}$ " required />
+                        <label for="mail"></label>
+                        <input type="email" id="mail" name="user_email" placeholder="Email address" class="contact-input" pattern="[a-z0-9._%+-]+@[a-z0-9.-]+\.[a-z]{2,4}$" required />
                     </li>
-                    <li class="text ">
-                        <label for="msg "></label>
-                        <textarea id="msg " name="user_message " rows="5 " placeholder="Enter text here (max 500 characters) " class="textarea " maxlength="500 " required></textarea>
+                    <li class="text">
+                        <label for="msg"></label>
+                        <textarea id="msg" name="user_message" rows="5" placeholder="Enter text here (max 500 characters)" class="textarea" maxlength="500" required></textarea>
                     </li>
                 </ul>
-                <button type='submit' class="b ">Get in touch</button>
+                <button type='submit' class="b">Get in touch</button>
 
             </form>
 
         </div>
-        <div class="contact-form-icons ">
-            <ul class="box3 ">
-                <li class="social-media ">
-                    <a href=" "><img src="./images/twitter.png " alt="twitter " /></a>
+        <div class="contact-form-icons">
+            <ul class="box3">
+                <li class="social-media">
+                    <a href=" "><img src="./images/twitter.png" alt="twitter" /></a>
+                </li>
+                <li class="social-media">
+                    <a href=" "><img src="./images/Linkedin.png" alt="Linkedln" /></a>
                 </li>
                 <li class="social-media ">
-                    <a href=" "><img src="./images/Linkedin.png " alt="Linkedln " /></a>
+                    <a href=" "><img src="./images/github.png" alt="m" /></a>
                 </li>
-                <li class="social-media ">
-                    <a href=" "><img src="./images/github.png " alt="m " /></a>
+                <li class="social-media">
+                    <a href=" "><img src="./images/fingers.png " alt="github" /></a>
                 </li>
-                <li class="social-media ">
-                    <a href=" "><img src="./images/fingers.png " alt="github " /></a>
-                </li>
-                <li class="social-media ">
-                    <a href=" "><img src="./images/m.png " alt="two fingers " /></a>
+                <li class="social-media">
+                    <a href=" "><img src="./images/m.png" alt="two fingers" /></a>
                 </li>
             </ul>
 
         </div>
-        <a class="my-email-address " href="mailto:ciraganenicole@gmail.com ">ciraganenicole@gmail.com</a>
+        <a class="my-email-address" href="mailto:ciraganenicole@gmail.com">ciraganenicole@gmail.com</a>
     </footer>
-    <script type="module" src="./src/script.js"></script>
+    <script type='module' src="./source/script.js"></script>
 </body>
 
 </html>
