+++ source/data.js
@@ -1,0 +1,87 @@
+const projet1 = {
+  imageSrc: './images/popup-desk.png',
+  titre: 'Multi-Post Stories Gain+Glory',
+  technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+  id: 'firstProject',
+  popupDescription: {
+    imageSrc: './images/popup-desk.png',
+    titre: 'Multi-Post Stories Gain+Glory',
+    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+    description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    demoLink: 'https://ciraganenicole.github.io/Portfolio/',
+    source: 'https://github.com/ciraganenicole/Portfolio',
+  },
+};
+const projet2 = {
+  imageSrc: './images/popup-desk.png',
+  titre: 'Multi-Post Stories Gain+Glory',
+  technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+  id: 'secondProject',
+  popupDescription: {
+    imageSrc: './images/popup-desk.png',
+    titre: 'Multi-Post Stories Gain+Glory',
+    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+    description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    demoLink: 'https://ciraganenicole.github.io/Portfolio/',
+    source: 'https://github.com/ciraganenicole/Portfolio',
+  },
+};
+const projet3 = {
+  imageSrc: './images/popup-desk.png',
+  titre: 'Multi-Post Stories Gain+Glory',
+  technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+  id: 'thirdProject',
+  popupDescription: {
+    imageSrc: './images/popup-desk.png',
+    titre: 'Multi-Post Stories Gain+Glory',
+    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+    description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    demoLink: 'https://ciraganenicole.github.io/Portfolio/',
+    source: 'https://github.com/ciraganenicole/Portfolio',
+  },
+};
+const projet4 = {
+  imageSrc: './images/popup-desk.png',
+  titre: 'Multi-Post Stories Gain+Glory',
+  technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+  id: 'fourthProject',
+  popupDescription: {
+    imageSrc: './images/popup-desk.png',
+    titre: 'Multi-Post Stories Gain+Glory',
+    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+    description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    demoLink: 'https://ciraganenicole.github.io/Portfolio/',
+    source: 'https://github.com/ciraganenicole/Portfolio',
+  },
+};
+const projet5 = {
+  imageSrc: './images/popup-desk.png',
+  titre: 'Multi-Post Stories Gain+Glory',
+  technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+  id: 'fifthProject',
+  popupDescription: {
+    imageSrc: './images/popup-desk.png',
+    titre: 'Multi-Post Stories Gain+Glory',
+    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+    description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    demoLink: 'https://ciraganenicole.github.io/Portfolio/',
+    source: 'https://github.com/ciraganenicole/Portfolio',
+  },
+};
+const projet6 = {
+  imageSrc: './images/popup-desk.png',
+  titre: 'Multi-Post Stories Gain+Glory',
+  technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+  id: 'sixthProject',
+  popupDescription: {
+    imageSrc: './images/popup-desk.png',
+    titre: 'Multi-Post Stories Gain+Glory',
+    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
+    description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
+    demoLink: 'https://ciraganenicole.github.io/Portfolio/',
+    source: 'https://github.com/ciraganenicole/Portfolio',
+  },
+};
+export const works = [projet1, projet2, projet3, projet4, projet5, projet6];
+
+export default projet2;
