--- src/data.js
@@ -1,87 +1,0 @@
-const projet1 = {
-    imageSrc: './images/popup-img.png',
-    titre: 'Multi-Post Stories Gain+Glory',
-    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-    id: 'firstProject',
-    popupDescription: {
-        imageSrc: './images/popup-desk.png',
-        titre: 'Multi-Post Stories Gain+Glory',
-        technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-        description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
-        demoLink: 'https://lexoxo.github.io/Portfolio-setup-and-mobile-version-skeleton/',
-        source: 'https://github.com/lexoxo/Portfolio-setup-and-mobile-version-skeleton',
-    },
-};
-const projet2 = {
-    imageSrc: './images/popup-img.png',
-    titre: 'Multi-Post Stories Gain+Glory',
-    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-    id: 'secondProject',
-    popupDescription: {
-        imageSrc: './images/popup-desk.png',
-        titre: 'Multi-Post Stories Gain+Glory',
-        technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-        description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
-        demoLink: 'https://lexoxo.github.io/Portfolio-setup-and-mobile-version-skeleton/',
-        source: 'https://github.com/lexoxo/Portfolio-setup-and-mobile-version-skeleton',
-    },
-};
-const projet3 = {
-    imageSrc: './images/popup-img.png',
-    titre: 'Multi-Post Stories Gain+Glory',
-    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-    id: 'thirdProject',
-    popupDescription: {
-        imageSrc: './images/popup-desk.png',
-        titre: 'Multi-Post Stories Gain+Glory',
-        technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-        description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
-        demoLink: 'https://lexoxo.github.io/Portfolio-setup-and-mobile-version-skeleton/',
-        source: 'https://github.com/lexoxo/Portfolio-setup-and-mobile-version-skeleton',
-    },
-};
-const projet4 = {
-    imageSrc: './images/popup-img.png',
-    titre: 'Multi-Post Stories Gain+Glory',
-    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-    id: 'fourthProject',
-    popupDescription: {
-        imageSrc: './images/popup-desk.png',
-        titre: 'Multi-Post Stories Gain+Glory',
-        technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-        description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
-        demoLink: 'https://lexoxo.github.io/Portfolio-setup-and-mobile-version-skeleton/',
-        source: 'https://github.com/lexoxo/Portfolio-setup-and-mobile-version-skeleton',
-    },
-};
-const projet5 = {
-    imageSrc: './images/popup-img.png',
-    titre: 'Multi-Post Stories Gain+Glory',
-    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-    id: 'fifthProject',
-    popupDescription: {
-        imageSrc: './images/popup-desk.png',
-        titre: 'Multi-Post Stories Gain+Glory',
-        technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-        description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
-        demoLink: 'https://lexoxo.github.io/Portfolio-setup-and-mobile-version-skeleton/',
-        source: 'https://github.com/lexoxo/Portfolio-setup-and-mobile-version-skeleton',
-    },
-};
-const projet6 = {
-    imageSrc: './images/popup-img.png',
-    titre: 'Multi-Post Stories Gain+Glory',
-    technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-    id: 'sixthProject',
-    popupDescription: {
-        imageSrc: './images/popup-desk.png',
-        titre: 'Multi-Post Stories Gain+Glory',
-        technolgies: ['Ruby on rails', 'css', 'javascript', 'html'],
-        description: " Lorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the releaLorem Ipsum is simply dummy text of the printing and typesetting industry. Lorem Ipsum has been the industry's standard dummy text ever since the 1500s, when an unknown printer took a galley of type and scrambled it 1960s with the relea",
-        demoLink: 'https://lexoxo.github.io/Portfolio-setup-and-mobile-version-skeleton/',
-        source: 'https://github.com/lexoxo/Portfolio-setup-and-mobile-version-skeleton',
-    },
-};
-export const works = [projet1, projet2, projet3, projet4, projet5, projet6];
-
-export default projet2;
